% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{of_dyn}
\alias{of_dyn}
\title{Objective function for estimating dynamic GMWM parameters}
\usage{
of_dyn(theta, nu_hat, x, tau, Omega, WN, QN, AR1)
}
\arguments{
\item{theta}{A vector of parameter}

\item{nu_hat}{A matrix of wavelet variances}

\item{x}{A vector of data}

\item{tau}{A vector containing the scales}

\item{Omega}{A matrix of weights}

\item{WN}{A boolean inidicating whether a gaussian White Noise is 
included in the model}

\item{QN}{A boolean inidicating whether a Quantization Noise is included 
in the model}

\item{AR1}{A non-negative integer indicating the number of Autoregressive 
process(es) of order 1 to be included in the model}
}
\value{
Returns a double $$\frac{1}{2}\frac{1}{K}\sum_{k=1}^K\lVert\hat{\nu}_k - \nu(\theta,x_k)\rVert^2_{\Omega}$$
}
\description{
Objective function for estimating dynamic GMWM parameters
}
